pyproject.toml
[tool.poetry]
authors = ["Francisco Del Roio <francipvb@hotmail.com>"]
description = ""
name = "interview-evaluation"
version = "0.1.0"

[tool.poetry.dependencies]
fastapi = "^0.73.0"
pydantic = {extras = ["email", "dotenv"], version = "^1.9.0"}
python = "^3.9"
uvicorn = "^0.17.4"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
factory-boy = "^3.2.1"
flake8 = "^4.0.1"
ipython = "^8.0.1"
mypy = "^0.931"
pytest = "^7.0.1"
pytest-cov = "^3.0.0"
requests = "^2.27.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.mypy]
plugins = [
  "pydantic.mypy.plugin.main",
]

[tool.black]
line-length = 79
